Writing Flow: Always add an empty at the end of the block list
