+++ editor/edit-post/modes/visual-editor/style.scss
@@ -124,3 +124,14 @@
 		}
 	}
 }
+
+.editor-visual-editor .editor-default-block-appender {
+	max-width: $visual-editor-max-width + ( 2 * $block-mover-padding-visible );
+	clear: left;
+	margin: 0 auto;
+	position: relative;
+
+	@include break-small() {
+		padding: 0 ( $block-padding + $block-mover-padding-visible );
+	}
+}
